Moved NPCAction into own file
